move interaction types
