+++ packages/sdk/src/interaction/interaction.ts
@@ -1,47 +1,8 @@
 import {invariant} from "@onflow/util-invariant"
 
-export const UNKNOWN /*                       */ = "UNKNOWN"
-export const SCRIPT /*                        */ = "SCRIPT"
-export const TRANSACTION /*                   */ = "TRANSACTION"
-export const GET_TRANSACTION_STATUS /*        */ = "GET_TRANSACTION_STATUS"
-export const GET_ACCOUNT /*                   */ = "GET_ACCOUNT"
-export const GET_EVENTS /*                    */ = "GET_EVENTS"
-export const PING /*                          */ = "PING"
-export const GET_TRANSACTION /*               */ = "GET_TRANSACTION"
-export const GET_BLOCK /*                     */ = "GET_BLOCK"
-export const GET_BLOCK_HEADER /*              */ = "GET_BLOCK_HEADER"
-export const GET_COLLECTION /*                */ = "GET_COLLECTION"
-export const GET_NETWORK_PARAMETERS /*        */ = "GET_NETWORK_PARAMETERS"
+import { IAcct, ACCOUNT, PARAM, ARGUMENT, UNKNOWN, OK, IIx, AUTHORIZER, PAYER, SCRIPT, TRANSACTION, GET_TRANSACTION_STATUS, GET_TRANSACTION, GET_ACCOUNT, GET_EVENTS, PING, GET_BLOCK, GET_BLOCK_HEADER, GET_COLLECTION, GET_NETWORK_PARAMETERS, BAD, PROPOSER } from "@onflow/typedefs"; 
 
-export const BAD /* */ = "BAD"
-export const OK /*  */ = "OK"
-
-export const ACCOUNT /*  */ = "ACCOUNT"
-export const PARAM /*    */ = "PARAM"
-export const ARGUMENT /* */ = "ARGUMENT"
-
-export const AUTHORIZER /* */ = "authorizer"
-export const PAYER /*      */ = "payer"
-export const PROPOSER /*   */ = "proposer"
-
-export interface IAcct {
-  "kind": typeof ACCOUNT,
-  "tempId": string | null,
-  "addr": string | null,
-  "keyId": number | string | null,
-  "sequenceNum": number | null,
-  "signature": any | null,
-  "signingFunction": any | null,
-  "resolve": any | null,
-  "role": {
-    "proposer": boolean,
-    "authorizer": boolean,
-    "payer": boolean,
-    "param": boolean,
-  },
-  authorization: any,
-}
-type AcctFn = (IAcct) => IAcct;
+type AcctFn = (acct: IAcct) => IAcct;
 type IAcctFn = AcctFn & Partial<IAcct>;
 
 const ACCT = `{
@@ -80,49 +41,6 @@ const ARG = `{
   "resolve": null,
   "resolveArgument": null
 }`
-
-interface IIx {
-  "tag": string,
-  "assigns": Record<string, any>,
-  "status": string,
-  "reason": string | null,
-  "accounts": Record<string, any>,
-  "params": Record<string, any>,
-  "arguments": Record<string, any>,
-  "message": {
-    "cadence": string | null,
-    "refBlock": string | null,
-    "computeLimit": string | null,
-    "proposer": string | null,
-    "payer": string | null,
-    "authorizations": string[],
-    "params": Record<string, any>[],
-    "arguments": Record<string, any>[]
-  },
-  "proposer": string | null,
-  "authorizations": string[],
-  "payer": string[],
-  "events": {
-    "eventType": string | null,
-    "start": string | null,
-    "end": string | null,
-    "blockIds": string[]
-  },
-  "transaction": {
-    "id": string | null
-  },
-  "block": {
-    "id": string | null,
-    "height": string | null,
-    "isSealed": boolean | null
-  },
-  "account": {
-    "addr": string | null
-  },
-  "collection": {
-    "id": string | null
-  }
-}
 
 const IX = `{
   "tag":"${UNKNOWN}",
@@ -174,11 +92,11 @@ export const initInteraction = (): IIx => JSON.parse(IX)
 const CHARS = "abcdefghijklmnopqrstuvwxyz0123456789".split("")
 const randChar = () => CHARS[~~(Math.random() * CHARS.length)]
 export const uuid = () => Array.from({length: 10}, randChar).join("")
-export const isNumber = d => typeof d === "number"
-export const isArray = d => Array.isArray(d)
-export const isObj = d => d !== null && typeof d === "object"
-export const isNull = d => d == null
-export const isFn = d => typeof d === "function"
+export const isNumber = (d: any): d is number => typeof d === "number"
+export const isArray = (d: any): d is any[] => Array.isArray(d)
+export const isObj = (d: any): d is Record<string, any> => d !== null && typeof d === "object"
+export const isNull = (d: any): d is null => d == null
+export const isFn = (d: any): d is Function => typeof d === "function"
 
 export const isInteraction = (ix: IIx) => {
   if (!isObj(ix) || isNull(ix) || isNumber(ix)) return false
@@ -191,7 +109,7 @@ export const Ok = (ix: IIx) => {
   return ix
 }
 
-export const Bad = (ix: IIx, reason) => {
+export const Bad = (ix: IIx, reason: string) => {
   ix.status = BAD
   ix.reason = reason
   return ix
@@ -227,7 +145,7 @@ export const prepAccount = (acct: IAcct | IAcctFn, opts: IPrepAccountOpts = {rol
   invariant(opts.role != null, "Account must have a role")
 
   const ACCOUNT = initAccount()
-  const role = opts.role
+  const role = opts.role as (typeof AUTHORIZER | typeof PAYER | typeof PROPOSER)
   const tempId = uuid()
 
   if (acct.authorization && isFn(acct.authorization))
@@ -236,7 +154,7 @@ export const prepAccount = (acct: IAcct | IAcctFn, opts: IPrepAccountOpts = {rol
 
   const resolve = acct.resolve
   if (resolve)
-    acct.resolve = (acct, ...rest) =>
+    acct.resolve = (acct: IAcct, ...rest: any[]) =>
       [resolve, prepAccountKeyId].reduce(
         async (d, fn) => fn(await d, ...rest),
         acct
@@ -265,7 +183,7 @@ export const prepAccount = (acct: IAcct | IAcctFn, opts: IPrepAccountOpts = {rol
   return ix
 }
 
-export const makeArgument = arg => ix => {
+export const makeArgument = (arg: Record<string, any>) => (ix: IIx)  => {
   let tempId = uuid()
   ix.message.arguments.push(tempId)
 
@@ -295,7 +213,7 @@ export const makeGetBlockHeader /*          */ = makeIx(GET_BLOCK_HEADER)
 export const makeGetCollection /*           */ = makeIx(GET_COLLECTION)
 export const makeGetNetworkParameters /*    */ = makeIx(GET_NETWORK_PARAMETERS)
 
-const is = wat => (ix: IIx) => ix.tag === wat
+const is = (wat: string) => (ix: IIx) => ix.tag === wat
 
 export const isUnknown /*                 */ = is(UNKNOWN)
 export const isScript /*                  */ = is(SCRIPT)
@@ -314,9 +232,9 @@ export const isOk /*  */ = (ix: IIx) => ix.status === OK
 export const isBad /* */ = (ix: IIx) => ix.status === BAD
 export const why /*   */ = (ix: IIx) => ix.reason
 
-export const isAccount /*  */ = account => account.kind === ACCOUNT
-export const isParam /*    */ = param => param.kind === PARAM
-export const isArgument /* */ = argument => argument.kind === ARGUMENT
+export const isAccount /*  */ = (account: Record<string, any>) => account.kind === ACCOUNT
+export const isParam /*    */ = (param: Record<string, any>) => param.kind === PARAM
+export const isArgument /* */ = (argument: Record<string, any>) => argument.kind === ARGUMENT
 
 const hardMode = (ix: IIx) => {
   for (let key of Object.keys(ix)) {
@@ -326,7 +244,7 @@ const hardMode = (ix: IIx) => {
   return ix
 }
 
-const recPipe = async (ix: IIx, fns = []) => {
+const recPipe = async (ix: IIx, fns: (Function | IIx)[] = []): Promise<any> => {
   try {
     ix = hardMode(await ix)
     if (isBad(ix)) throw new Error(`Interaction Error: ${ix.reason}`)
@@ -342,19 +260,19 @@ const recPipe = async (ix: IIx, fns = []) => {
   }
 }
 
-export const pipe = (...args) => {
+export const pipe = (...args: any[]) => {
   const [arg1, arg2] = args
-  if (isArray(arg1) && arg2 == null) return d => pipe(d, arg1)
+  if (isArray(arg1) && arg2 == null) return (d: any) => pipe(d, arg1)
   return recPipe(arg1, arg2)
 }
 
-const identity = (v, ..._) => v
+const identity = (v: any, ..._: any[]) => v
 
-export const get = (ix: IIx, key: string, fallback) => {
+export const get = (ix: IIx, key: string, fallback: any) => {
   return ix.assigns[key] == null ? fallback : ix.assigns[key]
 }
 
-export const put = (key: string, value) => (ix: IIx) => {
+export const put = (key: string, value: any) => (ix: IIx) => {
   ix.assigns[key] = value
   return Ok(ix)
 }
