+++ packages/typedefs/src/index.ts
@@ -288,3 +288,5 @@ export type Provider = {
    */
   name: string
 }
+
+export * from "./interaction"
